+++ StockAutomation/BusinessLayer/Services/SubscriberService.cs
@@ -6,6 +6,7 @@ using Microsoft.EntityFrameworkCore;
 
 namespace BusinessLayer.Services;
 
+using System.Net;
 using System.Net.Mail;
 
 public class SubscriberService : ISubscriberService
@@ -64,6 +65,14 @@ public class SubscriberService : ISubscriberService
     public async Task<Result<bool, Error>> DeleteSubscribersAsync(List<int> ids)
     {
         var subscribers = await _context.Subscribers.Where(s => ids.Contains(s.Id)).ToListAsync();
+        if (subscribers.Count != ids.Count)
+        {
+            return new Error
+            {
+                ErrorType = ErrorType.SubscribersNotFound,
+                Message = "One or more subscribers could not be found or do not exist. No action performed."
+            };
+        }
 
         _context.Subscribers.RemoveRange(subscribers);
         await _context.SaveChangesAsync();
